+++ src/App/Container/Hooks/useText.js
@@ -1,5 +1,5 @@
-import  { useState } from "react";
-import { initialText } from "./InitialText.js";
+import { useState } from 'react';
+import { initialText } from './InitialText.js';
 const useText = (initialValue = initialText) => {
   const [state, setState] = useState(initialValue);
   return [state, setState];